fix: prevent a package removal to fail and lock it on the git database
This is a long standing bug and particularly annoying when you have a package that builds its
dependencies using golang and set their dependencies as readonly. Because the build is not cleaned
afterwards, whenever you decide to remove that package, `rm` will fail because those files are
defined as readonly, thus can't be removed.

This fixes that by ensuring that the package to be removed has the correct permissions before
issuing the `rm`.