fix: enhance our built version detector
Previously, as of v2, we started detecting whether a package version would differ than its built
one. This is particular useful for vcs packages, which aren't required to update the `pkgver` field
every time a new version is pushed out to VCS, instead, makepkg relies on its `pkgver()` function to
find what's the version it will have. Now, obviously, because `pkgver()` is a function and it runs
after the sources are fetched, extracted and `prepared()`'d, much can happen in this process,
including pre-build processes that we have no control over. Due to that, we need to make sure that
during the version verification process, we install the build dependencies so `pkgver()` can run
properly and therefore we can safely identify its build version by looking at the modified PKGBUILD.

This might hit the performance when checking for version on `VCS` packages, but we can try and
improve that in the future.